fixing some syntaxe error

+++ num_system.c
@@ -145,7 +145,7 @@ int print_hex_X(va_list arg)
 int print_pointer(va_list arg)
 {
 	void *ptr = va_arg(arg, void *);
-	unsigned long int num = (unsigned long long int)ptr;
+	unsigned long long int num = (unsigned long long int)ptr;
 	unsigned long long int t;
 	char hex_d[] = "0123456789abcdef";
 	char digit;
